requirements.txt
Flask==3.0.3
mysql-connector==2.2.9